Script to generate intermediate binary files
This script reads the .gif files from the radar station and converts
them to an intermediate binary format.  In this format, pixels with no
rain have a value of 0, and pixels with rain have positive integer
values that become larger as the intensity of the rain increases.
There is a header block with details of the file, and a data block.
The data block is gzip compressed.