Update the intermediate binary file format
Decided that I want to record total rain pixel weight in the
intermediate binary file.  Updated the version number of the file, and
adjusted it accordingly.  Also provided a new method for reading only
the header of a binary file, to avoid forcing a gzip on the contents
when we're not interested in the data itself.