Examine the distribution for thinning data
We want to remove some training elements that do not predict rain any
time in the following five hours.  It seems a good way to do this
might be to drop candidate elements that have the lowest summed
intensity of rain in the historical interval, as they just predict
that no rain leads to ongoing no rain.  This script generates data to
evaluate the safety of that approach, and produces a veto list we can
use to drop candidates.