Accomodate change in data format
We now have one more bit, indicating whether there is rain happening
at the end of the historical interval.  I'm less interested in the
question of whether it will rain in 10 minutes given that it's raining
now, and more curious about whether it will start or stop raining.

The generator needs to know how to read this new format.