Add hash protection and holdout data
We're about to start doing comparisons and changes to see how they
affect the training.  I don't want to find out later that I
accidentally changed my input set partway through, invalidating the
results.  So, I compute a hash on the training inputs, and make sure
it's the expected value.

I also added holdout data for evaluating the suitability of the
model.  This is a data set that is not seen during training or
validation, but is used after the training is completed.