Resolve I/J ROW/COL inconsistency
When we started with .gif files, we decided to use the internal data
layout, in which the data is laid out in row major order.  Our arrays,
therefore, are indexed by row,col, not I,J.  This confusion led to
some places where indices were reversed.  So, checked all the files
and cleaned them up appropriately.