Better handling of phantom rain
The phantom rain handling hasn't been very good, this means that the
rain predictor gets confused by very light rain falling close to the
radar facility.  Added another data element for the inner rings that
holds the fraction of rain pixels in the module that have rain pixels
in all four of their bordering pixels.  This number should be small
for phantom rain, and large for true light rainfall.