Big changes
With the larger data set, reading in for every experiment was taking
too long, so I set up to save the input numpy arrays to disc, now it
takes seconds, instead of half an hour.

Added support for the month variable, but it's not hooked in yet.

Added some dropout and normalization layers, and played with the
geometries.

Fixed a bug in the holdout reporting.